requirements.txt
# FastAPI & Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# Testing
pytest==8.3.4
requests==2.32.3

# PII Detection & NLP
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.8.3

# Data Processing
pandas==2.2.3
python-dotenv==1.0.1

# Database
psycopg2-binary==2.9.10

# Workflow Orchestration
apache-airflow==2.10.4

# OpenAI (for embeddings)
openai==1.59.5

# Vector Database
pinecone-client==5.0.1

# Caching
redis==5.2.0

# Jupyter Notebooks
jupyter==1.1.1
notebook==7.3.2
ipykernel==6.29.5
